refactor Company to SDL
