+++ packages/server/graphql/public/types/User.ts
@@ -1,0 +1,14 @@
+import {isSuperUser} from '../../../utils/authorization'
+import getDomainFromEmail from '../../../utils/getDomainFromEmail'
+import isCompanyDomain from '../../../utils/isCompanyDomain'
+import {UserResolvers} from '../resolverTypes'
+
+const User: UserResolvers = {
+  company: async ({email}, _args, {authToken}) => {
+    const domain = getDomainFromEmail(email)
+    if (!domain || !isCompanyDomain(domain) || !isSuperUser(authToken)) return null
+    return {id: domain}
+  }
+}
+
+export default User
